Tree finished
